// frontend/src/api/models/SubtitleCapitalizationMethod.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SubtitleCapitalizationMethod = {
    Default: 'default',
    Uppercase: 'uppercase',
    Lowercase: 'lowercase',
    CapitalizeFirstCharInWords: 'capitalize_first_char_in_words'
} as const;
export type SubtitleCapitalizationMethod = typeof SubtitleCapitalizationMethod[keyof typeof SubtitleCapitalizationMethod];


export function instanceOfSubtitleCapitalizationMethod(value: any): boolean {
    for (const key in SubtitleCapitalizationMethod) {
        if (Object.prototype.hasOwnProperty.call(SubtitleCapitalizationMethod, key)) {
            if (SubtitleCapitalizationMethod[key as keyof typeof SubtitleCapitalizationMethod] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SubtitleCapitalizationMethodFromJSON(json: any): SubtitleCapitalizationMethod {
    return SubtitleCapitalizationMethodFromJSONTyped(json, false);
}

export function SubtitleCapitalizationMethodFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitleCapitalizationMethod {
    return json as SubtitleCapitalizationMethod;
}

export function SubtitleCapitalizationMethodToJSON(value?: SubtitleCapitalizationMethod | null): any {
    return value as any;
}

export function SubtitleCapitalizationMethodToJSONTyped(value: any, ignoreDiscriminator: boolean): SubtitleCapitalizationMethod {
    return value as SubtitleCapitalizationMethod;
}