// frontend/src/api/models/LanguageCode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const LanguageCode = {
    AutoDetect: 'auto-detect',
    En: 'en',
    Tr: 'tr',
    Es: 'es',
    Fr: 'fr',
    De: 'de',
    It: 'it',
    Pt: 'pt',
    Ru: 'ru',
    Ja: 'ja',
    Ko: 'ko'
} as const;
export type LanguageCode = typeof LanguageCode[keyof typeof LanguageCode];


export function instanceOfLanguageCode(value: any): boolean {
    for (const key in LanguageCode) {
        if (Object.prototype.hasOwnProperty.call(LanguageCode, key)) {
            if (LanguageCode[key as keyof typeof LanguageCode] === value) {
                return true;
            }
        }
    }
    return false;
}

export function LanguageCodeFromJSON(json: any): LanguageCode {
    return LanguageCodeFromJSONTyped(json, false);
}

export function LanguageCodeFromJSONTyped(json: any, ignoreDiscriminator: boolean): LanguageCode {
    return json as LanguageCode;
}

export function LanguageCodeToJSON(value?: LanguageCode | null): any {
    return value as any;
}

export function LanguageCodeToJSONTyped(value: any, ignoreDiscriminator: boolean): LanguageCode {
    return value as LanguageCode;
}