// frontend/src/api/models/SubtitleApplication.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SubtitleConfiguration } from './SubtitleConfiguration';
import {
    SubtitleConfigurationFromJSON,
    SubtitleConfigurationFromJSONTyped,
    SubtitleConfigurationToJSON,
    SubtitleConfigurationToJSONTyped,
} from './SubtitleConfiguration';

/**
 * 
 * @export
 * @interface SubtitleApplication
 */
export interface SubtitleApplication {
    /**
     * 
     * @type {Array<SubtitleConfiguration>}
     * @memberof SubtitleApplication
     */
    subtitleConfiguration?: Array<SubtitleConfiguration>;
}

/**
 * Check if a given object implements the SubtitleApplication interface.
 */
export function instanceOfSubtitleApplication(value: object): value is SubtitleApplication {
    return true;
}

export function SubtitleApplicationFromJSON(json: any): SubtitleApplication {
    return SubtitleApplicationFromJSONTyped(json, false);
}

export function SubtitleApplicationFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitleApplication {
    if (json == null) {
        return json;
    }
    return {
        
        'subtitleConfiguration': json['subtitle_configuration'] == null ? undefined : ((json['subtitle_configuration'] as Array<any>).map(SubtitleConfigurationFromJSON)),
    };
}

export function SubtitleApplicationToJSON(json: any): SubtitleApplication {
    return SubtitleApplicationToJSONTyped(json, false);
}

export function SubtitleApplicationToJSONTyped(value?: SubtitleApplication | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'subtitle_configuration': value['subtitleConfiguration'] == null ? undefined : ((value['subtitleConfiguration'] as Array<any>).map(SubtitleConfigurationToJSON)),
    };
}